Refactor theme and import systems, improve module organization
