+++ test/test_theme_registration.py
@@ -6,7 +6,7 @@ import os
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 
 import ttkbootstrap as tb
-from core.custom_themes import register_custom_themes
+from core.theme import register_custom_themes
 from core.theme.user_themes import USER_THEMES
 
 def test_theme_registration():
