+++ test/test_theme_system.py
@@ -14,8 +14,8 @@ logging.basicConfig(
 )
 
 # Import our modules
-from core.theme_manager import ThemeManager
-from core.custom_themes import register_custom_themes
+from core.theme.theme_system import ThemeManager
+from core.theme import register_custom_themes
 from utils.data.settings_manager import load_user_theme, save_user_theme
 
 def test_theme_system():
